code/audio-text-aligner/requirements.txt
aeneas
chardet
Flask
flask-restplus
Flask-Uploads
python-magic
